# Alerting: Improve performance of /api/prometheus for large numbers of alerts.