The AlertsByImportance comparison function is a big cost in service API requests, especially for large numbers of alert instances. This is because on every comparison, it makes a slice of strings and sorts them. This change uses the Prometheus Labels type, which is an already sorted slice of name/values pairs. This makes comparison much faster.

This should not change the content of the API response - the `Labels` type has a custom marshal function that acts the exact same way as marshaling `map[string]string`.

Benchmark before and after:

`go test -v -run=^# -bench BenchmarkSortAlertsByImportance -count 5 -topk sort`

```
name                                         old time/op  new time/op  delta
SortAlertsByImportance/n_1000_k_16-32        15.6ms ± 2%   0.6ms ± 1%  -96.06%  (p=0.008 n=5+5)
SortAlertsByImportance/n_1000_k_100-32       15.6ms ± 1%   0.6ms ± 2%  -96.03%  (p=0.008 n=5+5)
SortAlertsByImportance/n_1000_k_1000-32      15.8ms ± 2%   0.6ms ± 1%  -96.10%  (p=0.008 n=5+5)
SortAlertsByImportance/n_1000_k_100000-32    15.5ms ± 1%   0.6ms ± 1%  -96.03%  (p=0.008 n=5+5)
SortAlertsByImportance/n_10000_k_16-32        159ms ± 1%     8ms ± 2%  -95.11%  (p=0.008 n=5+5)
SortAlertsByImportance/n_10000_k_100-32       158ms ± 1%     8ms ± 3%  -95.00%  (p=0.008 n=5+5)
SortAlertsByImportance/n_10000_k_1000-32      159ms ± 1%     8ms ± 4%  -94.87%  (p=0.008 n=5+5)
SortAlertsByImportance/n_10000_k_100000-32    158ms ± 0%     8ms ± 4%  -94.93%  (p=0.016 n=4+5)
SortAlertsByImportance/n_100000_k_16-32       1.67s ± 1%   0.14s ± 2%  -91.80%  (p=0.008 n=5+5)
SortAlertsByImportance/n_100000_k_100-32      1.67s ± 1%   0.14s ± 6%  -91.64%  (p=0.008 n=5+5)
SortAlertsByImportance/n_100000_k_1000-32     1.66s ± 1%   0.14s ± 4%  -91.72%  (p=0.008 n=5+5)
SortAlertsByImportance/n_100000_k_100000-32   1.66s ± 1%   0.14s ± 2%  -91.79%  (p=0.008 n=5+5)
```